refactor with functional components and hooks
