--- src/components/clipboard-inspector/file-info.js
@@ -1,14 +1,0 @@
-import * as React from "react";
-
-const FileInfo = (file) => {
-  return file
-    ? {
-        name: file.name,
-        size: file.size,
-        type: file.type,
-        url: URL.createObjectURL(file),
-      }
-    : null;
-};
-
-export default FileInfo;
